+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Rational
 
+## Unreleased (2022-11-24)
+
+- [feat]: `.add()` method for vector addition;
+
 ## 0.1.0 (2022-11-23)
 
 - [chore]: First release, just parsing user input functionality;